+++ cogstat/cogstat_dialogs.py
@@ -23,10 +23,10 @@ QtCore.QCoreApplication.translate = _gui
 
 def open_data_file():
     return str(QtWidgets.QFileDialog.getOpenFileName(None, _('Open data file'), '',
-                                                     '%s *.csv *.sav *.zsav *.por *.sas7bdat;;'
+                                                     '%s *.csv *.sav *.zsav *.por *.sas7bdat *.xpt;;'
                                                      '%s *.csv;;'
                                                      '%s *.sav *.zsav *.por;;'
-                                                     '%s *.sas7bdat'%
+                                                     '%s *.sas7bdat *.xpt'%
                                                      (_('All importable data files'), _('Text files'),
                                                       _('SPSS data files'), _('SAS data files'))
                                                      )[0])
@@ -36,10 +36,10 @@ def open_data_file():
 def open_demo_data_file():
     return str(QtWidgets.QFileDialog.getOpenFileName(None, _('Open data file'), os.path.dirname(csc.__file__) +
                                                      '/sample_data',
-                                                     '%s *.csv *.sav *.zsav *.por *.sas7bdat;;'
+                                                     '%s *.csv *.sav *.zsav *.por *.sas7bdat *.xpt;;'
                                                      '%s *.csv;;'
                                                      '%s *.sav *.zsav *.por;;'
-                                                     '%s *.sas7bdat'%
+                                                     '%s *.sas7bdat *.xpt'%
                                                      (_('All importable data files'), _('Text files'),
                                                       _('SPSS data files'), _('SAS data files'))
                                                      )[0])  #*.txt *.log *.tsv